♻️ Refactor: 체험 등록 관련 타입을 도메인 내부로 이동
